requirements.txt
flask==3.0.2
Werkzeug==3.1.3
Jinja2==3.1.3
gunicorn==22.0.0
yfinance==0.2.42
requests==2.31.0
python-dotenv==1.0.0 